fix account token positions count
